BAU: Move SelfServiceError to a single file
