--- express/src/lib/SelfServiceError/ErrorOptions.ts
@@ -1,7 +1,0 @@
-export default interface ErrorOptions {
-    template?: string;
-    values?: {[name: string]: string};
-    errorMessages?: {[name: string]: string};
-}
-
-export {ErrorOptions};
