--- express/src/lib/SelfServiceError/SelfServiceError.ts
@@ -1,12 +1,0 @@
-import ErrorOptions from "./ErrorOptions";
-
-class SelfServiceError extends Error {
-    options?: ErrorOptions;
-    constructor(message: string, options?: ErrorOptions) {
-        super(message);
-        this.name = "SelfServiceError";
-        this.options = options;
-    }
-}
-
-export {SelfServiceError};
