--- express/src/lib/SelfServiceError/index.ts
@@ -1,5 +1,0 @@
-import {ErrorOptions} from "./ErrorOptions";
-import {SelfServiceError} from "./SelfServiceError";
-
-export {SelfServiceError};
-export {ErrorOptions};
