+++ express/src/middleware/passwordValidator.ts
@@ -1,5 +1,5 @@
 import {NextFunction, Request, Response} from "express";
-import {SelfServiceError} from "../lib/SelfServiceError";
+import SelfServiceError from "../lib/self-service-error";
 
 type MiddlewareFunction<T, U, V> = (T: Request, U: Response, V: NextFunction) => void;
